Make certificate installation conditional

+++ src/services/certificate.ts
@@ -1,5 +1,6 @@
 import { execFile } from "child_process";
 import path from "path";
+import fs from "fs";
 // @ts-ignore
 import log from "electron-log";
 import util from "util";
@@ -12,18 +13,21 @@ const options = {
 const CONF_DIR = path.join(__dirname, "../extra/nginx/conf")
 
 export const addCertificate = async (): Promise<void> => {
-  log.info("Adding certificate")
+  log.info("Adding certificate if needed")
   return new Promise((resolve, reject) => {
-    sudo.exec(`"${CONF_DIR}/mkcert.exe" -install -key-file "${CONF_DIR}/key.pem" -cert-file "${CONF_DIR}/cert.pem" kh.ssl.ak.tiles.virtualearth.net khstorelive.azureedge.net *.virtualearth.net *.azureedge.net`, options,
-      function(error: Error, stdout: string, stderr: string) {
-        if (error) {
-          reject(error);
-          return;
+    if (fs.existsSync(`${CONF_DIR}/key.pem`) && fs.existsSync(`${CONF_DIR}/cert.pem`)) {
+      resolve();
+    } else {
+      sudo.exec(`"${CONF_DIR}/mkcert.exe" -install -key-file "${CONF_DIR}/key.pem" -cert-file "${CONF_DIR}/cert.pem" kh.ssl.ak.tiles.virtualearth.net khstorelive.azureedge.net *.virtualearth.net *.azureedge.net`, options,
+        function(error: Error, stdout: string, stderr: string) {
+          if (error) {
+            reject(error);
+            return;
+          }
+          log.info("Added certificate", stdout);
+          resolve();
         }
-        log.info("Added certificate", stdout);
-        resolve();
-      }
-    );
+      );
+    }
   })
-  
 };
